loadJSFunctions() transmet les erreurs de validation => déplacement des fonctions TypeScript dans le package lib/engine/js, pour permettre de générer le nouveau jsFunctions.go malgré les erreurs temporaires de validation Golang
